Update vendored jbig2dec 0.18 -> 0.20
Part of updating mupdf and its dependencies to 1.28.0-rc2 (jbig2dec is a
mupdf thirdparty submodule, commit dc15c39). The /* SumatraPDF markers in
jbig2_huffman.c and jbig2_symbol_dict.c are carried upstream by Artifex
(present identically in upstream), so taking upstream preserves them; no
SumatraPDF-specific local edits needed hand-merging. Per-file line endings
preserved. No files added/removed.

+++ ext/jbig2dec/jbig2_huffman.h
@@ -64,7 +64,7 @@ int jbig2_huffman_skip(Jbig2HuffmanState *hs);
 
 int jbig2_huffman_advance(Jbig2HuffmanState *hs, size_t advance);
 
-uint32_t jbig2_huffman_offset(Jbig2HuffmanState *hs);
+size_t jbig2_huffman_offset(Jbig2HuffmanState *hs);
 
 int32_t jbig2_huffman_get(Jbig2HuffmanState *hs, const Jbig2HuffmanTable *table, bool *oob);
